refactor: rename sendAnywhere -> send-anywhere
- for consistency with other example contracts
