+++ packages/orchestration/src/examples/README.md
@@ -3,7 +3,7 @@
 This directory contains sample contracts showcasing the Orchestration API. Each example demonstrates specific functionalities:
 
 - **basic-flows.contract.js**: Account creation and query sending
-- **sendAnywhere.contract.js**: Token sending across supported blockchains
+- **send-anywhere.contract.js**: Token sending across supported blockchains
 - **auto-stake-it.contract.js**: Automatic remote staking of received tokens
 
 ## In Progress
