+++ packages/orchestration/src/examples/send-anywhere.contract.js
@@ -3,7 +3,7 @@ import { AmountShape } from '@agoric/ertp';
 import { InvitationShape } from '@agoric/zoe/src/typeGuards.js';
 import { M } from '@endo/patterns';
 import { withOrchestration } from '../utils/start-helper.js';
-import * as flows from './sendAnywhere.flows.js';
+import * as flows from './send-anywhere.flows.js';
 import { prepareChainHubAdmin } from '../exos/chain-hub-admin.js';
 
 /**
