+++ packages/orchestration/test/examples/snapshots/send-anywhere.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/examples/sendAnywhere.test.ts`
+# Snapshot report for `test/examples/send-anywhere.test.ts`
 
-The actual snapshot is saved in `sendAnywhere.test.ts.snap`.
+The actual snapshot is saved in `send-anywhere.test.ts.snap`.
 
 Generated by [AVA](https://avajs.dev).
